Add reverse-engineering docs and parity fixes
Large update adding reverse-engineering reconstruction notes and completing multiple qagame parity tasks. IMPLEMENTATION_PLAN.md was expanded with Tasks A71–A86 documenting recovered qagame command/table wiring, many g_* CVar restorations, and verification steps. Numerous docs were added under docs/reverse-engineering (detailed qagame reconstructions and many steam-mapping round notes) and supporting docs updated. Source changes span cgame, client, game, renderer, platform, and tests to restore retail behavior and wiring; references/analysis and HLIL mappings were refreshed. Test suites and parity sentinels were updated/added and targeted pytest verifications are recorded in the plan.

+++ src/code/client/snd_mem.c
@@ -108,7 +108,7 @@ void SND_setup() {
 
 	cv = Cvar_Get( "com_soundMegs", DEF_COMSOUNDMEGS, CVAR_LATCH | CVAR_ARCHIVE );
 
-	scs = (cv->integer*1536);
+	scs = (cv->integer*1024);
 
 	buffer = malloc(scs*sizeof(sndBuffer) );
 	// allocate the stack based hunk allocator
